Rename to avoid redundancy with the folder name
