+++ snipssonos/services/device/discovery_service.py
