+++ snipssonos/services/node/music_playback_service.py
@@ -1,7 +1,7 @@
 import requests
 import logging
 
-from snipssonos.services.music.music_playback_service import MusicPlaybackService
+from snipssonos.services.music.playback_service import MusicPlaybackService
 
 class NodeMusicPlaybackService(MusicPlaybackService):  # TODO : Refactor this in next iteration ...
